画像一覧表示の機能追加
データベースは一旦保留にした
画像一覧表示ができるようになったため, APIのテストが可能であると思う

+++ app/src/main/java/com/example/androidapitest/Picture.kt
@@ -4,9 +4,11 @@ import io.realm.RealmObject
 import io.realm.annotations.PrimaryKey
 import java.util.*
 
+// Don't use this version
+
 open class Picture : RealmObject() {
     @PrimaryKey
-    var id: Int = 0
+    var id: Long = 0
     var date: Date = Date()
     var name: String = ""
     var picture: String = ""
